requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
python-dotenv==1.0.1
sse-starlette==2.1.0
tenacity==8.3.0
rich==13.7.1
python-multipart==0.0.9
